move await value test util
